+++ src/x509/clu_x509_sign.c
@@ -1026,7 +1026,7 @@ int wolfCLU_CertSetDate(WOLFSSL_X509* x509, int days)
         }
 
         notBefore = wolfSSL_ASN1_TIME_adj(NULL, t, 0, 0);
-        notAfter = wolfSSL_ASN1_TIME_adj(NULL, t, days, 0);
+        notAfter = wolfSSL_ASN1_TIME_adj(NULL, t, (int)days, 0);
         if (notBefore == NULL || notAfter == NULL) {
             wolfCLU_LogError("Error creating not before/after dates");
             ret = WOLFCLU_FATAL_ERROR;
